requirements.txt
cbpro==1.1.4
certifi==2020.6.20
chardet==3.0.4
finnhub-python==2.2.0
idna==2.10
Jinja2==2.11.2
MarkupSafe==1.1.1
oauthlib==3.1.0
pymongo==3.5.1
PySocks==1.7.1
python-dotenv==0.14.0
requests==2.24.0
requests-oauthlib==1.3.0
six==1.10.0
sortedcontainers==2.2.2
SQLAlchemy==1.3.19
tweepy==3.9.0
urllib3==1.25.10
websocket-client==0.40.0
